Route back detailed movie
RESULTADO:  {
  "id": 123,
  "genres": [
    "Adventure",
    "Animation",
    "Fantasy"
  ],
  "title": "The Lord of the Rings",
  "description": "The Fellowship of the Ring embark on a journey to destroy the One Ring and end Sauron's reign over Middle-earth.",
  "rating": 6.565,
  "user_reviews": 674,
  "release_date": "November 14, 1978",
  "duration": "132 minutes.",
  "poster": "https://image.tmdb.org/t/p/w400/liW0mjvTyLs7UCumaHhx3PpU4VT.jpg",
  "back_poster": "https://image.tmdb.org/t/p/w400/jOuCWdh0BE6XPu2Vpjl08wDAeFz.jpg",
  "trailer": "https://www.youtube.com/watch?v=9qhL2_UxXM0"
}
