+++ api/src/Routes/index.js
@@ -3,14 +3,23 @@ const router = Router();
 // const { Movies, Genres } = require('../db.js');
 
 // Import functions from controllers:
-const { getMoviesByIdApi } = require('../controllers API/index.js');
+const {
+  getMoviesByIdApi,
+  getTrailerMovie,
+} = require('../controllers API/detailedMovie.js');
 
 // ROUTES:
-// GET VIDEOGAME BY ID:
+// Get movie from API by ID with trailer:
 router.get('/movies/:id', async (req, res) => {
   try {
     const { id } = req.params;
-    const movieDetail = await getMoviesByIdApi(id);
+    let movieDetail = await getMoviesByIdApi(id);
+    const trailer = await getTrailerMovie(id);
+
+    movieDetail = {
+      ...movieDetail,
+      trailer,
+    };
     res.send(movieDetail);
   } catch (error) {
     return res.status(404).send(error);
